Revert "Fix T62436: New added Torus won't appear in Local View"
This reverts commit 9054b39b27c4198f436329f78af43fe3f3175508.

Using the region data to find the screen & space-data doesn't make sense
since local-view is space-data there is no reason to pass in a region
argument, then lookup the space-data.

There was no way to remove an object or check it's local view state
as there was in 2.7x.

This can be made closer to 2.7x API and simplified.